Moving nghttp's numeric_name function to util.cc and using this to generate the address name in HttpServer.cc
