+++ src/util.h
@@ -29,6 +29,7 @@
 
 #include <unistd.h>
 #include <getopt.h>
+#include <netdb.h>
 
 #include <cstring>
 #include <cassert>
@@ -416,6 +417,8 @@ void write_uri_field(std::ostream &o, const char *uri, const http_parser_url &u,
                      http_parser_url_fields field);
 
 bool numeric_host(const char *hostname);
+
+std::string numeric_name(addrinfo *addr);
 
 // Opens |path| with O_APPEND enabled.  If file does not exist, it is
 // created first.  This function returns file descriptor referring the
